# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PRS API",
        "description": "API documentation for the PRS application.",
        "contact": {
            "email": "support@prs.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/sales-manager/dashboard": {
            "get": {
                "tags": [
                    "Sales Manager Dashboard"
                ],
                "summary": "Get Sales Manager dashboard summary data",
                "operationId": "9cd3f98ba0b4e55a4036d0960e60fc92",
                "responses": {
                    "200": {
                        "description": "Dashboard data for the logged-in Sales Manager",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "retailer_order_stats": {
                                            "type": "object"
                                        },
                                        "distributor_order_stats": {
                                            "type": "object"
                                        },
                                        "counts": {
                                            "type": "object"
                                        },
                                        "top_fieldstaff": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sales-manager/pending-retailers": {
            "get": {
                "tags": [
                    "Sales Manager Dashboard"
                ],
                "summary": "List retailers waiting for approval",
                "operationId": "ecd8d0c4cbbefa66594c0afabf58a0a9",
                "responses": {
                    "200": {
                        "description": "List of pending retailers"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sales-manager/retailers/{id}/approve": {
            "post": {
                "tags": [
                    "Sales Manager Dashboard"
                ],
                "summary": "Approve (activate) a retailer account",
                "operationId": "477a26f1ba120e3aa322348febfe30f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retailer approved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sales-manager/fieldstaffs": {
            "get": {
                "tags": [
                    "Sales Manager Dashboard"
                ],
                "summary": "List all field staff assigned to this Sales Manager",
                "operationId": "5a015b37e508bec3756d034f46e0ea05",
                "responses": {
                    "200": {
                        "description": "List of field staff"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sales-manager/retailers": {
            "get": {
                "tags": [
                    "Sales Manager Dashboard"
                ],
                "summary": "List all retailers under this Sales Manager (via field staff)",
                "operationId": "8b695aacfa79b1ea2969f3503b8c4b3c",
                "responses": {
                    "200": {
                        "description": "List of retailers"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sales-manager/retailers/loyalty-points": {
            "get": {
                "tags": [
                    "Sales Manager Dashboard"
                ],
                "summary": "List all retailers under this Sales Manager with their loyalty points",
                "operationId": "3d73d79e4b5ca87f526e1d617a98c476",
                "responses": {
                    "200": {
                        "description": "List of retailers' loyalty points"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sales-manager/retailers/{id}/loyalty-points": {
            "get": {
                "tags": [
                    "Sales Manager Dashboard"
                ],
                "summary": "Get loyalty points summary and history for a retailer",
                "operationId": "ab4fe7e4cf9cce3feb4b4997df35a1ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Loyalty points data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sales-manager/retailer-orders": {
            "get": {
                "tags": [
                    "Sales Manager Dashboard"
                ],
                "summary": "List all retailer orders under this Sales Manager",
                "operationId": "6ec61769605dd12574e08adde5618fa6",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "processing",
                                "accepted",
                                "delivered",
                                "cancelled",
                                "rejected"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of retailer orders"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sales-manager/retailer-orders/{id}/update-status": {
            "post": {
                "tags": [
                    "Sales Manager Dashboard"
                ],
                "summary": "Accept or reject a pending retailer order",
                "description": "Sales managers can accept a pending order (moves to processing) or reject it.",
                "operationId": "21827c29f262d47f6d26b543c7145883",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "processing",
                                            "rejected"
                                        ]
                                    },
                                    "cancellation_reason": {
                                        "description": "Required if status is rejected",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order status updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sales-manager/distributor-insights": {
            "get": {
                "tags": [
                    "Sales Manager Dashboard"
                ],
                "summary": "List all distributors with summary stats for Sales Manager",
                "operationId": "a5c72f9858abf7177e2dee175b4ff3cb",
                "responses": {
                    "200": {
                        "description": "List of distributors with insights"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}